requirements.txt
Django==3.2.19
djangorestframework==3.14.0
Pillow==9.5.0
